Add logos in HTML Utils
