+++ packages/browser/examples/MockUp.html
@@ -18,7 +18,8 @@
         './assets/config/layer/3DTiles.json',
         './assets/config/layer/elevation.json',
         './assets/config/layer/geoJSONs.json',
-                './assets/config/layer/base_maps.json',
+        './assets/config/layer/base_maps.json',
+        './assets/config/widget/about.json'
       ]).then((configs) => {
         // http://proj4js.org/
         // define a projection as a string and reference it that way
@@ -39,7 +40,7 @@
 
         const app = new udvizBrowser.MockUp(
           extent,
-          configs['frame3D_planars'][0]
+          configs
         );
 
         const frame3DPlanar = app.getFrame3DPlanar();
